drivers: npmx: shell: move errlog code to the separate file
