+++ drivers/npmx/CMakeLists.txt
@@ -50,6 +50,7 @@ if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(shell/adc.c)
     zephyr_library_sources(shell/buck.c)
     zephyr_library_sources(shell/charger.c)
+    zephyr_library_sources(shell/errlog.c)
 endif()
 
 # Get npmx include file
