+++ drivers/npmx/npmx_shell.c
@@ -110,72 +110,6 @@ static npmx_vbusin_t *vbusin_instance_get(const struct shell *shell)
 	return npmx_instance ? npmx_vbusin_get(npmx_instance, 0) : NULL;
 }
 
-static const char *shell_err_to_field(npmx_callback_type_t type, uint8_t bit)
-{
-	static const char * err_fieldnames[][8] =
-	{
-		[NPMX_CALLBACK_TYPE_RSTCAUSE] =
-		{
-			[0] = "SHIPMODEEXIT",    [1] = "BOOTMONITORTIMEOUT",
-			[2] = "WATCHDOGTIMEOUT", [3] = "LONGPRESSTIMEOUT",
-			[4] = "THERMALSHUTDOWN", [5] = "VSYSLOW",
-			[6] = "SWRESET",
-		},
-		[NPMX_CALLBACK_TYPE_CHARGER_ERROR] =
-		{
-			[0] = "NTCSENSORERR",  [1] = "VBATSENSORERR",
-			[2] = "VBATLOW",       [3] = "VTRICKLE",
-			[4] = "MEASTIMEOUT",   [5] = "CHARGETIMEOUT",
-			[6] = "TRICKLETIMEOUT",
-		},
-		[NPMX_CALLBACK_TYPE_SENSOR_ERROR] =
-		{
-			[0] = "SENSORNTCCOLD",  [1] = "SENSORNTCCOOL",
-			[2] = "SENSORNTCWARM",  [3] = "SENSORNTCHOT",
-			[4] = "SENSORVTERM",    [5] = "SENSORRECHARGE",
-			[6] = "SENSORVTRICKLE", [7] = "SENSORVBATLOW",
-		},
-	};
-
-	return err_fieldnames[type][bit];
-}
-
-static void print_errlog(npmx_instance_t *p_pm, npmx_callback_type_t type, uint8_t mask)
-{
-	const struct shell *shell = (struct shell *)npmx_core_context_get(p_pm);
-
-	shell_print(shell, "%s:", npmx_callback_to_str(type));
-	for (uint8_t i = 0; i < 8; i++) {
-		if ((1U << i) & mask) {
-			shell_print(shell, "\t%s", shell_err_to_field(type, i));
-		}
-	}
-}
-
-static int cmd_errlog_get(const struct shell *shell, size_t argc, char **argv)
-{
-	ARG_UNUSED(argc);
-	ARG_UNUSED(argv);
-
-	npmx_instance_t *npmx_instance = npmx_driver_instance_get(pmic_dev);
-	if (npmx_instance == NULL) {
-		return 0;
-	}
-
-	npmx_core_context_set(npmx_instance, (void *)shell);
-
-	npmx_core_register_cb(npmx_instance, print_errlog, NPMX_CALLBACK_TYPE_RSTCAUSE);
-	npmx_core_register_cb(npmx_instance, print_errlog, NPMX_CALLBACK_TYPE_CHARGER_ERROR);
-	npmx_core_register_cb(npmx_instance, print_errlog, NPMX_CALLBACK_TYPE_SENSOR_ERROR);
-
-	npmx_errlog_t *errlog_instance = npmx_errlog_get(npmx_instance, 0);
-	npmx_error_t err_code = npmx_errlog_reset_errors_check(errlog_instance);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "error log");
-	}
-	return 0;
-}
-
 static npmx_gpio_mode_t gpio_mode_convert(const struct shell *shell, uint32_t mode)
 {
 	switch (mode) {
@@ -1799,10 +1733,6 @@ static int cmd_vbusin_status_connected_get(const struct shell *shell, size_t arg
 	return 0;
 }
 
-/* Creating subcommands (level 2 command) array for command "errlog". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_errlog, SHELL_CMD(get, NULL, "Get error logs", cmd_errlog_get),
-			       SHELL_SUBCMD_SET_END);
-
 /* Creating subcommands (level 4 command) array for command "gpio config debounce". */
 SHELL_STATIC_SUBCMD_SET_CREATE(sub_gpio_config_debounce,
 			       SHELL_CMD(set, NULL, "Set debounce status",
@@ -2099,7 +2029,6 @@ SHELL_STATIC_SUBCMD_SET_CREATE(
 	sub_vbusin, SHELL_CMD(current_limit, &sub_vbusin_current_limit, "Current limit", NULL),
 	SHELL_CMD(status, &sub_vbusin_status, "Status", NULL), SHELL_SUBCMD_SET_END);
 
-SHELL_SUBCMD_ADD((npmx), errlog, &sub_errlog, "Reset errors logs", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), gpio, &sub_gpio, "GPIO", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), ldsw, &sub_ldsw, "LDSW", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), led, &sub_led, "LED", NULL, 1, 0);
